requirements.txt
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1